+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.2] - 2023-01-25
+### Changed
+- [cdk-stacks] - update dependencies
+- [agent-app] - update dependencies
+- [demo-website] - update dependencies
+- [cdk-stacks] - move lambdas from NODE 12 to 14
+- [README.md] - adding in instructions for updating IAM role to allow it to assume itself 
+
 ## [2.3.1] - 2022-11-23
 ### Changed
 - [cdk-stacks] - update dependencies